+++ packages/gamut-styles/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.1"></a>
+## [2.0.1](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut-styles@2.0.0...@codecademy/gamut-styles@2.0.1) (2018-06-12)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
+
+
+
+
+<a name="2.0.0"></a>
+# [2.0.0](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut-styles@1.3.25...@codecademy/gamut-styles@2.0.0) (2018-06-11)
+
+
+### Features
+
+* Merge standard & core variables ([#123](https://github.com/RyzacInc/gamut/issues/123)) ([5bf24b0](https://github.com/RyzacInc/gamut/commit/5bf24b0))
+
+
+### BREAKING CHANGES
+
+* `gamut-styles/variables` renamed to `gamut-styles/utils/variables`
+
+
+
+
+
 <a name="1.3.25"></a>
 ## [1.3.25](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut-styles@1.3.24...@codecademy/gamut-styles@1.3.25) (2018-05-23)